+++ ydb/core/tx/columnshard/engines/changes/general_compaction.cpp
@@ -1,4 +1,5 @@
 #include "general_compaction.h"
+#include "merge_subset.h"
 
 #include "compaction/merger.h"
 #include "counters/general.h"
@@ -34,129 +35,6 @@ std::vector<TWritePortionInfoWithBlobsResult> TGeneralCompactColumnEngineChanges
     }
     return result;
 }
-
-class ISubsetToMerge {
-private:
-    virtual std::vector<TPortionToMerge> DoBuildPortionsToMerge(const TConstructionContext& context, const std::set<ui32>& seqDataColumnIds,
-        const std::shared_ptr<TFilteredSnapshotSchema>& resultFiltered, const THashSet<ui64>& usedPortionIds) const = 0;
-
-public:
-    virtual ~ISubsetToMerge() = default;
-    std::vector<TPortionToMerge> BuildPortionsToMerge(const TConstructionContext& context, const std::set<ui32>& seqDataColumnIds,
-        const std::shared_ptr<TFilteredSnapshotSchema>& resultFiltered, const THashSet<ui64>& usedPortionIds) const {
-        return DoBuildPortionsToMerge(context, seqDataColumnIds, resultFiltered, usedPortionIds);
-    }
-    virtual ui64 GetColumnMaxChunkMemory() const = 0;
-};
-
-class TReadPortionToMerge: public ISubsetToMerge {
-private:
-    TReadPortionInfoWithBlobs ReadPortion;
-    const std::shared_ptr<TGranuleMeta> GranuleMeta;
-
-    std::shared_ptr<NArrow::TColumnFilter> BuildPortionFilter(const std::optional<NKikimr::NOlap::TGranuleShardingInfo>& shardingActual,
-        const std::shared_ptr<NArrow::TGeneralContainer>& batch, const TPortionInfo& pInfo, const THashSet<ui64>& portionsInUsage) const {
-        std::shared_ptr<NArrow::TColumnFilter> filter;
-        if (shardingActual && pInfo.NeedShardingFilter(*shardingActual)) {
-            std::set<std::string> fieldNames;
-            for (auto&& i : shardingActual->GetShardingInfo()->GetColumnNames()) {
-                fieldNames.emplace(i);
-            }
-            auto table = batch->BuildTableVerified(fieldNames);
-            AFL_VERIFY(table);
-            filter = shardingActual->GetShardingInfo()->GetFilter(table);
-        }
-        NArrow::TColumnFilter filterDeleted = NArrow::TColumnFilter::BuildAllowFilter();
-        if (pInfo.GetMeta().GetDeletionsCount()) {
-            if (pInfo.HasInsertWriteId()) {
-                AFL_VERIFY(pInfo.GetMeta().GetDeletionsCount() == pInfo.GetRecordsCount());
-                filterDeleted = NArrow::TColumnFilter::BuildDenyFilter();
-            } else {
-                auto table = batch->BuildTableVerified(std::set<std::string>({ TIndexInfo::SPEC_COL_DELETE_FLAG }));
-                AFL_VERIFY(table);
-                auto col = table->GetColumnByName(TIndexInfo::SPEC_COL_DELETE_FLAG);
-                AFL_VERIFY(col);
-                AFL_VERIFY(col->type()->id() == arrow::Type::BOOL);
-                for (auto&& c : col->chunks()) {
-                    auto bCol = static_pointer_cast<arrow::BooleanArray>(c);
-                    for (ui32 i = 0; i < bCol->length(); ++i) {
-                        filterDeleted.Add(!bCol->GetView(i));
-                    }
-                }
-            }
-            if (GranuleMeta->GetPortionsIndex().HasOlderIntervals(pInfo, portionsInUsage)) {
-                filterDeleted = NArrow::TColumnFilter::BuildAllowFilter();
-            }
-        }
-        if (filter) {
-            *filter = filter->And(filterDeleted);
-        } else if (!filterDeleted.IsTotalAllowFilter()) {
-            filter = std::make_shared<NArrow::TColumnFilter>(std::move(filterDeleted));
-        }
-        return filter;
-    }
-
-    virtual std::vector<TPortionToMerge> DoBuildPortionsToMerge(const TConstructionContext& context, const std::set<ui32>& seqDataColumnIds,
-        const std::shared_ptr<TFilteredSnapshotSchema>& resultFiltered, const THashSet<ui64>& usedPortionIds) const override {
-        auto blobsSchema = ReadPortion.GetPortionInfo().GetSchema(context.SchemaVersions);
-        auto batch = ReadPortion.RestoreBatch(*blobsSchema, *resultFiltered, seqDataColumnIds, false).DetachResult();
-        auto shardingActual = context.SchemaVersions.GetShardingInfoActual(GranuleMeta->GetPathId());
-        std::shared_ptr<NArrow::TColumnFilter> filter = BuildPortionFilter(shardingActual, batch, ReadPortion.GetPortionInfo(), usedPortionIds);
-        return { TPortionToMerge(batch, filter) };
-    }
-
-public:
-    TReadPortionToMerge(TReadPortionInfoWithBlobs&& rPortion, const std::shared_ptr<TGranuleMeta>& granuleMeta)
-        : ReadPortion(std::move(rPortion))
-        , GranuleMeta(granuleMeta) {
-    }
-
-    virtual ui64 GetColumnMaxChunkMemory() const override {
-        ui64 result = 0;
-        for (auto&& i : ReadPortion.GetChunks()) {
-            result = std::max<ui64>(result, i.second->GetRawBytesVerified());
-        }
-        return result;
-    }
-};
-
-class TWritePortionsToMerge: public ISubsetToMerge {
-private:
-    std::vector<TWritePortionInfoWithBlobsResult> WritePortions;
-
-    virtual std::vector<TPortionToMerge> DoBuildPortionsToMerge(const TConstructionContext& context, const std::set<ui32>& seqDataColumnIds,
-        const std::shared_ptr<TFilteredSnapshotSchema>& resultFiltered, const THashSet<ui64>& /*usedPortionIds*/) const override {
-        std::vector<TPortionToMerge> result;
-        for (auto&& i : WritePortions) {
-            auto blobsSchema = i.GetPortionResult().GetPortionInfo().GetSchema(context.SchemaVersions);
-            auto batch = i.RestoreBatch(*blobsSchema, *resultFiltered, seqDataColumnIds, false).DetachResult();
-            result.emplace_back(TPortionToMerge(batch, nullptr));
-        }
-        return result;
-    }
-
-    virtual ui64 GetColumnMaxChunkMemory() const override {
-        ui64 result = 0;
-        for (auto&& wPortion : WritePortions) {
-            for (auto&& i : wPortion.GetPortionResult().GetRecordsVerified()) {
-                result = std::max<ui64>(result, i.GetMeta().GetRawBytes());
-            }
-        }
-        return result;
-    }
-
-public:
-    TWritePortionsToMerge(std::vector<TWritePortionInfoWithBlobsResult>&& portions)
-        : WritePortions(std::move(portions)) {
-        ui32 idx = 0;
-        for (auto&& i : WritePortions) {
-            i.GetPortionConstructor().MutablePortionConstructor().SetPortionId(++idx);
-            i.GetPortionConstructor().MutablePortionConstructor().MutableMeta().SetCompactionLevel(0);
-            i.RegisterFakeBlobIds();
-            i.FinalizePortionConstructor();
-        }
-    }
-};
 
 TConclusionStatus TGeneralCompactColumnEngineChanges::DoConstructBlobs(TConstructionContext& context) noexcept {
     TSimplePortionsGroupInfo insertedPortions;
